create custom hook for build collection
